Extract util.ts from index.ts
内部から `index.ts` 内の関数 `sleep` などを呼び出すと循環参照になってしまうため、`util.ts` に切り出しました。
